Back 버튼 링크 변경, style 컴포넌트 분리
