+++ src/AllActivityPage.jsx
@@ -1,6 +1,6 @@
 import React from 'react';
 
-import { Link } from 'react-router-dom';
+import { useHistory, Link } from 'react-router-dom';
 
 import resultData from './resultData';
 
@@ -9,16 +9,7 @@ import { vitaminsImages } from './assets/images';
 import styled from '@emotion/styled';
 
 import Title from './styles/Title';
-
-const Back = styled.div({
-  position: 'fixed',
-  right: '10px',
-  '& a': {
-    textDecoration: 'none',
-    fontSize: '18px',
-    color: '#F2C94C',
-  }
-})
+import Back from './styles/Back';
 
 const List = styled.ul({
   listStyle: 'none',
@@ -55,9 +46,14 @@ const ResultItem = styled.p({
 });
 
 export default function AllActivityPage() {
+  const history = useHistory();
+
+  function handleClick() {
+    return history.goBack();
+  }
   return (
     <div>
-      <Back><Link to="/result">Back</Link></Back>
+      <Back><button type="button" onClick={handleClick}>Back</button></Back>
       <Title>All Vitamins</Title>
       <List>
         {resultData.map(({ text: { results } }) => (
